Renamed the 3 transforms by removing the suffix Field in the name of the components
