+++ src/js/components/pipelines/transforms/Filter.jsx
